Create spoofMAC.sh
A simple script to spoof your MAC address. Put it in a cronjob to run @reboot if you always want it spoofed. To do this, you will have to alter the code from asking for a wireless interface to writing in the specific one you would like to spoof on reboot.